+++ tests/test_callbacks.py
@@ -14,7 +14,7 @@ def test_raw_stat_callback(tmp_path):
         "--env",
         "CartPole-v1",
         "-params",
-        "callback:'rl_zoo.callbacks.RawStatisticsCallback'",
+        "callback:'rl_zoo3.callbacks.RawStatisticsCallback'",
         "--tensorboard-log",
         f"{tmp_path}",
     ]
@@ -32,7 +32,7 @@ def test_tqdm_callback(tmp_path):
         "--env",
         "CartPole-v1",
         "-params",
-        "callback:'rl_zoo.callbacks.TQDMCallback'",
+        "callback:'rl_zoo3.callbacks.TQDMCallback'",
         "--tensorboard-log",
         f"{tmp_path}",
     ]